+++ R/spinners.R
@@ -774,6 +774,38 @@ bs4_spinner <- function(
 }
 
 #' @rdname spinners
+#' @export 
+bs5_spinner <- function(
+  style = c("spin", "grow"),
+  color = c(
+    "primary",
+    "secondary",
+    "success",
+    "danger",
+    "warning",
+    "info",
+    "light",
+    "dark"
+  )
+) {
+  style <- match.arg(style)
+  color <- match.arg(color)
+
+  cl <- "border"
+  if(style == "grow")
+    cl <- "grow"
+
+  tags$div(
+    class = sprintf("spinner-%s text-%s", cl, color),
+    role = "status",
+    tags$span(
+      class = "visually-hidden",
+      "Loading..."
+    )
+  )
+}
+
+#' @rdname spinners
 #' @export
 spin_google <- function(){
   construct(div(class = "cm-spinner"), 7)