+++ NEWS.md
@@ -9,8 +9,8 @@ the attendant instead of `httr::progress`.
 - Fix major issue with `Hostess`, it was basically broken
 [#108](https://github.com/JohnCoene/waiter/issues/108).
 - Add `withWaiter` as requested in 
-[#105](https://github.com/JohnCoene/waiter/issues/105)
-- Added `bs4_spinner`.
+[#105](https://github.com/JohnCoene/waiter/issues/105).
+- Added `bs4_spinner` and `bs5_spinner`.
 
 # waiter 0.2.3